# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "1.0.0",
      "title": "Users API"
    },
    "servers": [
      {
        "url": "localhost:3000"
      }
    ],
    "paths": {
      "/users": {
        "get": {
          "summary": "Sample endpoint: Returns a list of users",
          "operationId": "listUser",
          "tags": [
            "user"
          ],
          "parameters": [
            {
              "name": "firstName",
              "in": "query",
              "description": "filter users list by first name",
              "required": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Sample response: list of users",
              "headers": {
                "x-requestId": {
                  "description": "Uniq request Id",
                  "schema": {
                    "type": "string"
                  }
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UsersList"
                  }
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        }
      },
      "/users/{userId}": {
        "get": {
          "summary": "Sample endpoint: Returns User by Id",
          "operationId": "User",
          "tags": [
            "user"
          ],
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "description": "Id of user to get",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Sample response: User by Id",
              "headers": {
                "x-requestId": {
                  "description": "Uniq request Id",
                  "schema": {
                    "type": "string"
                  }
                }
              },
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "required": [
            "id",
            "firstName"
          ],
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "firstName": {
              "type": "string"
            },
            "tag": {
              "type": "string"
            }
          }
        },
        "UsersList": {
          "type": "object",
          "properties": {
            "results": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/User"
              }
            },
            "count": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "Error": {
          "type": "object",
          "required": [
            "code",
            "message"
          ],
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "message": {
              "type": "string"
            }
          }
        }
      },
      "securitySchemes": {
        "BasicAuth": {
          "type": "http",
          "scheme": "basic"
        }
      }
    },
    "security": [
      {
        "BasicAuth": []
      }
    ]
  }